requirements.txt
streamlit
mysql-connector-python
pandas
Pillow
